# src/main/resources/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: AgendApp API
  version: 1.0.0
  description: API para gestión de turnos y reservas con pagos
servers:
  - url: http://localhost:8080/api
paths:
  /auth/register:
    post:
      summary: Registro de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario registrado
  /auth/login:
    post:
      summary: Login de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Token JWT
...
components:
  schemas:
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
...